feat: admin middleware

+++ routes/web.php
@@ -27,36 +27,35 @@ Route::get('/dashboard', function () {
     return view('dashboard');
 })->middleware(['auth', 'verified'])->name('dashboard');
 
-Route::middleware('auth')->group(function () {
-    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 
-    Route::get('/menu', [MenuController::class, 'index'])->name('menu');
-    Route::post('/menu', [MenuController::class, 'store']);
-    Route::get('/menu/create', [MenuController::class, 'create']);
-    Route::get('/menu/{id}', [MenuController::class, 'show'])->name('menu.show');
-    Route::get('/menu/{id}/edit', [MenuController::class, 'edit']);
-    Route::put('/menu/{id}', [MenuController::class, 'update']);
-    Route::delete('/menu/{id}', [MenuController::class, 'destroy']);
-    
-    Route::get('/type', [TypeController::class, 'index'])->name('type');
-    Route::post('/type', [TypeController::class, 'store']);
-    Route::get('/type/create', [TypeController::class, 'create']);
-    Route::get('/type/{id}/edit', [TypeController::class, 'edit']);
-    Route::put('/type/{id}', [TypeController::class, 'update']);
-    Route::delete('/type/{id}', [TypeController::class, 'destroy']);
+Route::get('/profile', [ProfileController::class, 'edit'])->middleware(['auth'])->name('profile.edit');
+Route::patch('/profile', [ProfileController::class, 'update'])->middleware(['auth'])->name('profile.update');
+Route::delete('/profile', [ProfileController::class, 'destroy'])->middleware(['auth'])->name('profile.destroy');
 
-    Route::post('/menu/{id}', [ReviewController::class,'store'])->name('review.create');
+Route::get('/menu', [MenuController::class, 'index'])->middleware(['auth'])->name('menu');
+Route::post('/menu', [MenuController::class, 'store'])->middleware(['auth']);
+Route::get('/menu/create', [MenuController::class, 'create'])->middleware(['auth', 'admin']);
+Route::get('/menu/{id}', [MenuController::class, 'show'])->middleware(['auth'])->name('menu.show');
+Route::get('/menu/{id}/edit', [MenuController::class, 'edit'])->middleware(['auth', 'admin']);
+Route::put('/menu/{id}', [MenuController::class, 'update'])->middleware(['auth']);
+Route::delete('/menu/{id}', [MenuController::class, 'destroy'])->middleware(['auth']);
 
-    Route::get('/cart', [CartController::class, 'index'])->name('cart');
-    Route::post('/menu', [CartController::class, 'store'])->name('cart.create');
-    Route::delete('/cart', [CartController::class, 'destroy'])->name('cart.delete');
-    Route::put('/cart', [CartController::class, 'update'])->name('cart.update');
+Route::get('/type', [TypeController::class, 'index'])->middleware(['auth'])->name('type');
+Route::post('/type', [TypeController::class, 'store'])->middleware(['auth']);
+Route::get('/type/create', [TypeController::class, 'create'])->middleware(['auth', 'admin']);
+Route::get('/type/{id}/edit', [TypeController::class, 'edit'])->middleware(['auth', 'admin']);
+Route::put('/type/{id}', [TypeController::class, 'update'])->middleware(['auth']);
+Route::delete('/type/{id}', [TypeController::class, 'destroy'])->middleware(['auth']);
 
-    Route::get('/order', [OrderController::class,'index'])->name('order');
-    Route::post('/order', [OrderController::class, 'store']);
-    Route::get('/order/create', [OrderController::class, 'create']);
-});
+Route::post('/menu/{id}', [ReviewController::class, 'store'])->middleware(['auth'])->name('review.create');
+
+Route::get('/cart', [CartController::class, 'index'])->middleware(['auth'])->name('cart');
+Route::post('/menu', [CartController::class, 'store'])->middleware(['auth'])->name('cart.create');
+Route::delete('/cart', [CartController::class, 'destroy'])->middleware(['auth'])->name('cart.delete');
+Route::put('/cart', [CartController::class, 'update'])->middleware(['auth'])->name('cart.update');
+
+Route::get('/order', [OrderController::class, 'index'])->middleware(['auth'])->name('order');
+Route::post('/order', [OrderController::class, 'store'])->middleware(['auth']);
+Route::get('/order/create', [OrderController::class, 'create'])->middleware(['auth']);
 
 require __DIR__ . '/auth.php';
